refactor: modularize tauri command domains (closes #108)
